Fixing auto-generated name suffix in case of already used name
